Refactor

+++ supervised_learning/error_analysis/0-create_confusion.py
@@ -22,12 +22,10 @@ def create_confusion_matrix(labels, logits):
     num_classes = labels.shape[1]
 
     # Initialize the confusion matrix
-    confusion_matrix = np.zeros((num_classes, num_classes),
-                                dtype=int)
+    confusion_matrix = np.zeros((num_classes, num_classes))
 
     # Populate the confusion matrix
     for true, pred in zip(true_labels, predicted_labels):
         confusion_matrix[true, pred] += 1
 
     return confusion_matrix
-
